refactor(Dropdown): move types to a separate file
